Break out context callback into utility files
